+++ cre.database.sh
@@ -25,9 +25,9 @@ cd $2
 prefix=`date +%Y-%m-%d`
 cat ~/cre/cre.database.header *.c4r > $prefix.c4r.sample_wise.csv
 cre.database_merge.py $prefix.c4r.sample_wise.csv $prefix.c4r.variant_wise.csv
-rm *.c4r
+#rm *.c4r
 
 #create files for report generation
-cat $prefix.c4r.variant_wise.csv | awk -F ',' '{print $1"-"$2"-"$3"\t"$12}'  > ~/Desktop/reference_tables/seen_in_c4r_counts.txt
-cat $prefix.c4r.variant_wise.csv | awk -F ',' '{print $1"-"$2"-"$3"\t"$13}'  > ~/Desktop/reference_tables/seen_in_c4r_samples.txt
+cat $prefix.c4r.variant_wise.csv | awk -F ',' '{print $1"-"$2"-"$3"\t"$(NF-1)}'  > ~/Desktop/reference_tables/seen_in_c4r_counts.txt
+cat $prefix.c4r.variant_wise.csv | awk -F ',' '{print $1"-"$2"-"$3"\t"$NF}'  > ~/Desktop/reference_tables/seen_in_c4r_samples.txt